requirements.txt
eth-brownie~=1.19.3
python-dotenv~=0.16.0
joblib~=1.2.0
click~=8.1.3
pandas~=1.5.2
psutil~=5.9.2
fastparquet~=2023.2.0
pyarrow~=11.0.0
imgkit~=1.2.3
pytest~=6.2.5
pytest-mock~=3.10.0
tabulate